// gestion-escolar-webapp/src/app/interfaces/usuarios-interfaces.ts
// Cree estas interfaces para mantener más limpio el codigo
// Tambien para dividirlo por tipo de usuario
// La ventaja es que se pueden importar y ya no se ve feo afuera de la clase

/// Interfaz datos de usuario de maestro
export interface DatosMaestro {
  id: number;
  id_trabajador: number;
  first_name: string;
  last_name: string;
  email: string;
  fecha_nacimiento: string;
  telefono: string;
  rfc: string;
  cubiculo: string;
  area_investigacion: number;
  materias_array: string[];
  campus:string;
  sueldo: number;
}

export interface DatosAlumno {
  id: number;
  matricula: string;
  first_name: string;
  last_name: string;
  email: string;
  fecha_nacimiento: string;
  telefono: string;
  rfc: string;
  curp: string;
  edad: number;
  ocupacion: string;  
  direccion: string;
  sexo: string; 
}


export interface DatosAdministrador {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  telefono: string;
  rfc: string;
  edad: number;
  ocupacion: string;
  clave_admin: string;
  contrato: String;
  grado_academico:String;
}